(#31) Add TwoPanel searchable, selectable history
Now is selectable and drawn similarly to the store window
