+++ Packages/Fluxity/Editor/ObjectWalker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc916172fc50acb4d9710f7a1c4cb19f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
